fix: CI clippy/test fixes + wave 3 modules (fuzz, bus, IPC, panels)
CI fixes:
- flight-ipc: manual_range_contains → (400..=600).contains()
- flight-metrics-http: approx_constant (3.14→3.125)
- flight-scheduler: collapsible_if + unnecessary_cast (Linux)
- flight-xplane: manual_is_multiple_of
- flight-hid: fd_safety_tests type inference (turbofish)
- flight-replay: useless_vec → array literals

Wave 3 modules:
- flight-bus: event router with priority routing
- flight-axis: pipeline stages (noise filter, response curve)
- flight-panels: protocol handler (Saitek multi-panel)
- flight-ipc: message types, connection pool, subscriptions
- Fuzz targets for DCS/HID/SimConnect/XPlane/IPC parsers

Test fixtures:
- flight-session: profile hierarchy test fixtures (global/msfs/C172)

+++ crates/flight-scheduler/src/unix.rs
@@ -825,7 +825,7 @@ impl LinuxTimerLoop {
     /// Utility method for time calculations.
     #[inline]
     pub fn timespec_to_nanos(ts: &libc::timespec) -> i64 {
-        ts.tv_sec as i64 * Self::NANOS_PER_SEC + ts.tv_nsec as i64
+        ts.tv_sec * Self::NANOS_PER_SEC + ts.tv_nsec
     }
 
     /// Calculate the difference between two timespecs in nanoseconds
